\name{giscolor-package}
\alias{giscolor-package}
\alias{giscolor}
\docType{package}
\title{
Add GRASS GIS color tables to R.
}
\description{
Adds more options for coloring grid-like images for display.
}
\details{
The color tables from GRASS GIS 7.0 have been made available as R palettes,
along with tools for adding palettes, displaying palettes, and using GIS
palettes to color R images.

Color tables do not have to be linear, and can be specified using percentages
or as data values.

The palettes included are adapted from the GRASS 7.0 color tables
\url{https://grasswiki.osgeo.org/wiki/Color_tableshttps://grasswiki.osgeo.org/wiki/Color_tables}.

\packageIndices{giscolor}
}
\author{
Sarah Goslee
Maintainer: Sarah Goslee <Sarah.Goslee@ars.usda.gov>
}
\keyword{ package }
